Rylan please work on this

+++ src/controllers/churchUserController.ts
@@ -1,8 +1,7 @@
 import { RequestHandler } from "express";
 import { ChurchUser } from "../models/churchUser";
-import { Church } from "../models/church";
-import { Event } from "../models/event";
-import { Op } from "sequelize";
+import { comparePasswords, hashPassword } from "../services/auth";
+import { signUserToken, verifyUser } from "../services/authService"
 
 
 
@@ -16,53 +15,59 @@ export const allUser: RequestHandler = async (req, res, next) => {
 
 export const createUser: RequestHandler = async (req, res, next) => {
     let newUser: ChurchUser = req.body;
-    if (
 
-        newUser.email,
-        newUser.password
+    // if (
+        
+    //   newUser.email,
+    //   newUser.password
 
-    ) {
-        let create = await ChurchUser.create(newUser);
-
-        res.status(201).json(create);
-    }
-    else {
-        res.status(400).send();
-    }
-
-    // if (newUser.email && newUser.password) {
-    //     // hassPass will go here 
-
-    //         let create = await ChurchUser.create(newUser);
-    //         res.status(200).json({
-    //             email: create.email,
-    //             passward: create.password,
-    //             userId: create.userId
-    //         });
-
-    // } else {
-    //     res.status(400).send('Email and password required.')
+    //   ) {
+    //     let create = await ChurchUser.create(newUser);
+        
+    //     res.status(201).json(create);
     // }
+    // else {
+    //     res.status(400).send();
+    // } 
+    
+    if (newUser.email && newUser.password) {
+        // hassPass will go here 
+    let hashedPasswrod = await hashPassword(newUser.password);
+    newUser.password =hashedPasswrod;
+      
+    let create = await ChurchUser.create(newUser);
+            res.status(200).json({
+                email: create.email,
+                userId: create.userId
+            });
+         
+    } else {
+        res.status(400).send('Email and password required.')
+    }
 }
 
 
 export const signInUser: RequestHandler = async (req, res, next) => {
 
-    let validUser: ChurchUser | null = await ChurchUser.findOne({
-        where: { email: req.body.email }
-    });
-    if (validUser) {
+
+let validUser: ChurchUser | null = await ChurchUser.findOne({
+    where: {email: req.body.email}
+});
+    if (validUser){
         // add authenication 
+        let matchPass = await comparePasswords(req.body.password, validUser.password);
+   
 
-        const matchPass = req.body.password === validUser.password;
-
+        // const matchPass = req.body.password === validUser.password;
+    
         if (matchPass) {
             // if pass matches, generate token
-            // let token = await tknSignUser(validUser);
-            res.status(200).json({ message: 'Authentication successful' });
+            let token = await signUserToken(validUser);
+            res.status(200).json({ token });
         } else {
             res.status(401).json('Password invalid');
-        }
+        } 
+
     } else {
         res.status(401).json('Email invalid')
     }
@@ -114,27 +119,59 @@ export const getUser: RequestHandler = async (req, res, next) => {
 
 
 
-export const modifyUser: RequestHandler = async (req, res, next) => {
-    let churchUser = req.params.id;
-    let newUser: ChurchUser = req.body;
-    let findUser = await ChurchUser.findByPk(churchUser);
-    if (findUser && findUser.userId === newUser.userId && newUser.email && newUser.password) {
-        await ChurchUser.update(newUser, { where: { userId: churchUser } });
+
+export const modifyUser: RequestHandler = async ( req, res, next ) => {
+    let newUser = req.body;
+    let user = await verifyUser(req);
+
+    //Does the user exist? if yes contiune
+    if (!user) {
+        return res.status(403).send();
+    }
+
+    let userId = parseInt(req.params.id);
+    newUser.userId = userId
+    newUser.password = await hashPassword(newUser.password)
+
+    //Is the user making the edit the same user editing themselves? if yes continue
+    if (user.userId != userId){
+        return res.status(403).send("Not the same user");
+    }
+
+    let foundUser = await ChurchUser.findByPk(userId);
+    if (!foundUser) {
+        return res.status(404).send();
+    }
+    if (foundUser.dataValues.userId === parseInt(newUser.userId)) {
+        await ChurchUser.update(newUser, {where: {userId}});
         res.status(200).json();
     }
     else {
-        res.status(400).json();
-    }
+        res.status(400).send();
 
 }
 
 
-export const deleteUser: RequestHandler = async (req, res, next) => {
-    let churchUser = req.params.id;
-    let findUser = await ChurchUser.findByPk(churchUser);
-    if (findUser) {
+
+export const deleteUser: RequestHandler = async ( req, res, next ) => {
+    let user = await verifyUser(req);
+
+    if (!user) {
+        return res.status(403).send();
+    }
+
+    let userId = parseInt(req.params.id);
+    let findUser = await ChurchUser.findByPk(userId);
+
+    //Is the user making the edit the same user editing themselves? if yes continue
+    if (user.userId != userId){
+        return res.status(403).send("Not the same user");
+    }
+
+    if (findUser){
         await ChurchUser.destroy({
-            where: { userId: churchUser }
+            where:{userId: userId}
+
         });
         res.status(200).json();
     }
